Feature/usuario (#14)
* Fix: Substituição do TipoUsuario por herança entre Usuario e Funcionario

* Feat: Métodos de listar clientes, administradores e estoquistas

* Fix: Troca do tipo dos IDs de int para Long

+++ src/main/java/br/edu/fatecgru/mercado_inteligente/controller/CategoriaController.java
@@ -30,7 +30,7 @@ public class CategoriaController {
 
 	// Busca por ID
 	@GetMapping("/{id}")
-	public Categoria buscarPorId(@PathVariable int id) {
+	public Categoria buscarPorId(@PathVariable Long id) {
 		return categoriaService.getById(id);
 	}
