+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/CategoriaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecgru.mercado_inteligente.model.entity.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 
 	// Método para fazer buscar por categoria
 	public List<Categoria> findByNomeContains(String nome);
