+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/ItemCarrinhoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecgru.mercado_inteligente.model.entity.ItemCarrinho;
 
-public interface ItemCarrinhoRepository extends JpaRepository<ItemCarrinho, Integer> {
+public interface ItemCarrinhoRepository extends JpaRepository<ItemCarrinho, Long> {
 
 }
