+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/MarcaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecgru.mercado_inteligente.model.entity.Marca;
 
-public interface MarcaRepository extends JpaRepository<Marca, Integer> {
+public interface MarcaRepository extends JpaRepository<Marca, Long> {
 
 	// Método para fazer buscar por categoria
 	public List<Marca> findByNomeContains(String nome);
