+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/MovimentacaoEstoqueRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecgru.mercado_inteligente.model.entity.MovimentacaoEstoque;
 
-public interface MovimentacaoEstoqueRepository extends JpaRepository<MovimentacaoEstoque, Integer> {
+public interface MovimentacaoEstoqueRepository extends JpaRepository<MovimentacaoEstoque, Long> {
 
 }
